requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
shapely>=2.1.0
pydantic==2.5.3
numpy>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

